requirements.txt
crewai>=0.80.0
google-genai>=1.0.0
python-dotenv>=1.0.0
rich>=13.0.0
